feat: split into server+shared packages
