+++ packages/server/package.json
@@ -1,0 +1,68 @@
+{
+  "name": "@tc39/tcq-server",
+  "version": "1.0.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/bterlson/tcq"
+  },
+  "description": "A TC39 Discussion Queue",
+  "main": "index.js",
+  "scripts": {
+    "build-frontend": "webpack --config ./src/client/webpack.dev.js",
+    "build-frontend-watch": "webpack --watch --config ./src/client/webpack.dev.js",
+    "build-backend": "tsc",
+    "build-production": "npm run build-backend && webpack -p --config ./src/client/webpack.prod.js",
+    "iterate": "npm run build-backend && concurrently \"npm run build-frontend-watch\" \"node ./dist/server/server.js\"",
+    "start": "node dist/server/server.js | \"./node_modules/.bin/bunyan\""
+  },
+  "author": "Brian Terlson",
+  "contributors": [
+    {
+      "name": "Daniel Rosenwasser",
+      "url": "https://github.com/DanielRosenwasser"
+    }
+  ],
+  "license": "Apache-2.0",
+  "engines": {
+    "node": "16.13.0"
+  },
+  "dependencies": {
+    "applicationinsights": "^1.0.0",
+    "axios": "^0.16.2",
+    "base64-url": "^2.0.0",
+    "body-parser": "^1.18.2",
+    "bunyan": "^1.8.12",
+    "connect-ensure-login": "^0.1.1",
+    "cookie-session": "^1.3.3",
+    "documentdb": "^1.12.2",
+    "documentdb-session": "^1.0.1",
+    "documentdb-typescript": "^1.0.6",
+    "express": "^4.15.4",
+    "express-bunyan-logger": "^1.3.2",
+    "express-session": "^1.15.5",
+    "github": "^11.0.0",
+    "p-limit": "^1.2.0",
+    "passport": "^0.4.0",
+    "passport-github2": "^0.1.12",
+    "socket.io": "^2.2.0",
+    "socket.io-client": "^2.0.3",
+    "uuid": "^3.1.0"
+  },
+  "devDependencies": {
+    "@types/axios": "^0.14.0",
+    "@types/body-parser": "^1.16.5",
+    "@types/bunyan": "^1.8.3",
+    "@types/connect-ensure-login": "^0.1.3",
+    "@types/documentdb": "^1.10.2",
+    "@types/documentdb-session": "^1.0.1",
+    "@types/express": "^4.0.39",
+    "@types/express-session": "^1.15.3",
+    "@types/node": "^8.0.28",
+    "@types/passport": "^0.3.4",
+    "@types/socket.io": "^1.4.31",
+    "@types/socket.io-client": "^1.4.31",
+    "@types/uuid": "^3.4.3",
+    "strict-event-emitter-types": "^1.2.0",
+    "typescript": "^2.8.1"
+  }
+}
