+++ packages/server/src/socket-hander.ts
@@ -1,14 +1,14 @@
-import Meeting from '../shared/Meeting';
-import Speaker from '../shared/Speaker';
-import AgendaItem from '../shared/AgendaItem';
+import Meeting from '../../shared/src/Meeting';
+import Speaker from '../../shared/src/Speaker';
+import AgendaItem from '../../shared/src/AgendaItem';
 import User, { getByUsername } from './User';
-import Reaction, { ReactionTypes } from '../shared/Reaction';
-import GitHubAuthenticatedUser from '../shared/GitHubAuthenticatedUser';
+import Reaction, { ReactionTypes } from '../../shared/src/Reaction';
+import GitHubAuthenticatedUser from '../../shared/src/GitHubAuthenticatedUser';
 import * as socketio from 'socket.io';
 import { isChair } from './User';
-import * as Message from '../shared/Messages';
+import * as Message from '../../shared/src/Messages';
 import { updateMeeting, getMeeting, getMeetingsCollection } from './db';
-import { TopicTypes } from '../shared/Speaker';
+import { TopicTypes } from '../../shared/src/Speaker';
 import gha from './ghapi';
 const PRIORITIES: Speaker['type'][] = ['poo', 'question', 'reply', 'topic'];
 import * as uuid from 'uuid';
