+++ packages/shared/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@tc39/tcq-shared",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "strict-event-emitter-types": "^1.2.0"
+  }
+}
