+++ packages/shared/src/AgendaItem.ts
