+++ packages/shared/src/GitHubAuthenticatedUser.ts
