+++ packages/shared/src/Meeting.ts
