+++ packages/shared/src/Speaker.ts
