+++ packages/shared/src/User.ts
