--- src/server/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../tsconfig.base.json",
-  "compilerOptions": {
-    "outDir": "../../dist",
-    "lib": ["es2017"],
-    "target": "es5",
-    "module": "commonjs"
-  }
-}
